+++ busltee/runner.go
@@ -12,7 +12,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/heroku/busl/busltee/buslteelogger"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 )
@@ -40,14 +39,14 @@ func Run(url string, args []string, conf *Config) (exitCode int) {
 	done := post(url, reader, conf)
 
 	if err := run(args, writer, writer); err != nil {
-		buslteelogger.WithFields(logrus.Fields{"count#busltee.exec.error": 1}).Error(err)
+		logWithFields(logrus.Fields{"count#busltee.exec.error": 1}).Error(err)
 		exitCode = exitStatus(err)
 	}
 
 	select {
 	case <-done:
 	case <-time.After(time.Second):
-		buslteelogger.WithFields(logrus.Fields{"count#busltee.exec.upload.timeout": 1}).Warn()
+		logWithFields(logrus.Fields{"count#busltee.exec.upload.timeout": 1}).Warn()
 	}
 
 	return exitCode
@@ -63,7 +62,7 @@ func setupLog(conf *Config) {
 }
 
 func monitor(subject string, ts time.Time) {
-	buslteelogger.WithFields(logrus.Fields{"time": time.Now().Sub(ts).Seconds()}).Warnf("%s.time", subject)
+	logWithFields(logrus.Fields{"time": time.Now().Sub(ts).Seconds()}).Warnf("%s.time", subject)
 }
 
 func post(url string, reader io.Reader, conf *Config) chan struct{} {
@@ -71,11 +70,11 @@ func post(url string, reader io.Reader, conf *Config) chan struct{} {
 
 	go func() {
 		if err := stream(url, reader, conf); err != nil {
-			buslteelogger.WithFields(logrus.Fields{"count#busltee.stream.error": 1}).Error(err)
+			logWithFields(logrus.Fields{"count#busltee.stream.error": 1}).Error(err)
 			// Prevent writes from blocking.
 			io.Copy(ioutil.Discard, reader)
 		} else {
-			buslteelogger.WithFields(logrus.Fields{"count#busltee.stream.success": 1}).Warn()
+			logWithFields(logrus.Fields{"count#busltee.stream.success": 1}).Warn()
 		}
 		close(done)
 	}()
@@ -88,7 +87,7 @@ func stream(url string, stdin io.Reader, conf *Config) (err error) {
 		if err = streamNoRetry(url, stdin, conf); !isTimeout(err) {
 			return err
 		}
-		buslteelogger.WithFields(logrus.Fields{"count#busltee.stream.retry": 1}).Warn()
+		logWithFields(logrus.Fields{"count#busltee.stream.retry": 1}).Warn()
 	}
 	return err
 }
@@ -99,7 +98,7 @@ func streamNoRetry(url string, stdin io.Reader, conf *Config) error {
 	defer monitor("busltee.stream", time.Now())
 
 	if url == "" {
-		buslteelogger.WithFields(logrus.Fields{"count#busltee.stream.missingurl": 1}).Warn()
+		logWithFields(logrus.Fields{"count#busltee.stream.missingurl": 1}).Warn()
 		return errMissingURL
 	}
 
@@ -228,7 +227,7 @@ func deliverSignals(cmd *exec.Cmd) {
 	go func() {
 		s := <-sigc
 
-		buslteelogger.WithFields(logrus.Fields{
+		logWithFields(logrus.Fields{
 			"signal": s,
 		}).Debug("received signal")
 
@@ -236,7 +235,7 @@ func deliverSignals(cmd *exec.Cmd) {
 		case syscall.SIGCHLD:
 		case syscall.SIGPIPE:
 		default:
-			buslteelogger.WithFields(logrus.Fields{"busltee.signal.deliver": s}).Info()
+			logWithFields(logrus.Fields{"busltee.signal.deliver": s}).Info()
 			cmd.Process.Signal(s)
 		}
 	}()
@@ -246,7 +245,7 @@ func wait(cmd *exec.Cmd) (*os.ProcessState, error) {
 	pgid, err := syscall.Getpgid(cmd.Process.Pid)
 	if err == nil {
 		defer func() {
-			buslteelogger.WithFields(logrus.Fields{
+			logWithFields(logrus.Fields{
 				"pgid": pgid,
 				"pid":  cmd.Process.Pid,
 			}).Debug("killing process group")
@@ -267,7 +266,7 @@ func exitStatus(err error) int {
 	err = errors.Cause(err)
 	if exit, ok := err.(*exec.ExitError); ok {
 		if status, ok := exit.Sys().(syscall.WaitStatus); ok {
-			buslteelogger.WithFields(logrus.Fields{
+			logWithFields(logrus.Fields{
 				"exit_status": status.ExitStatus(),
 				"pid":         exit.ProcessState.Pid(),
 				"signaled":    status.Signaled(),
